rename and add structure
