--- src/main/java/seedu/duke/command/Command.java
@@ -1,4 +1,0 @@
-package seedu.duke.command;
-
-public abstract class Command {
-}
