+++ src/main/java/seedu/recipeditor/Recipeditor.java
@@ -1,9 +1,13 @@
-package seedu.duke;
+package seedu.recipeditor;
 
-public class Duke {
+import seedu.recipeditor.ui.Ui;
+
+public class Recipeditor {
     /**
-     * Main entry-point for the java.duke.Duke application.
+     * Main entry point
      */
+
+
     public static void main(String[] args) {
         Ui.printGreeting();
         while (true) {
