+++ src/main/java/seedu/recipeditor/command/Command.java
@@ -1,0 +1,6 @@
+package seedu.recipeditor.command;
+
+public abstract class Command {
+    public String commandType = new String();
+    
+}
