+++ src/main/java/seedu/recipeditor/command/ListCommand.java
@@ -1,4 +1,4 @@
-package seedu.duke.command;
+package seedu.recipeditor.command;
 
 public class ListCommand extends Command {
 }
