+++ src/main/java/seedu/recipeditor/recipe/Ingredient.java
@@ -1,0 +1,5 @@
+package seedu.recipeditor.recipe;
+
+public class Ingredient {
+    
+}
