+++ src/main/java/seedu/recipeditor/storage/Storage.java
@@ -1,6 +1,8 @@
-package seedu.duke;
+package seedu.recipeditor.storage;
 
 import java.io.File;
+
+import seedu.recipeditor.ui.Ui;
 
 public class Storage {
     private static final String DATA_FILE_PATH = "data/data.txt";
